Closer to perfect gibs.

+++ src/game/default/g_ballistics.h
@@ -28,9 +28,9 @@
 void G_BlasterProjectile(g_edict_t *ent, const vec3_t start, const vec3_t dir,
 		int32_t speed, int16_t damage, int16_t knockback);
 void G_BulletProjectile(g_edict_t *ent, const vec3_t start, const vec3_t dir,
-		int16_t damage, int16_t knockback, int32_t hspread, int32_t vspread, int32_t mod);
+		int16_t damage, int16_t knockback, uint16_t hspread, uint16_t vspread, uint32_t mod);
 void G_ShotgunProjectiles(g_edict_t *ent, const vec3_t start, const vec3_t dir,
-		int16_t damage, int16_t knockback, int32_t hspread, int32_t vspread, int32_t count, int32_t mod);
+		int16_t damage, int16_t knockback, int32_t hspread, int32_t vspread, int32_t count, uint32_t mod);
 void G_HyperblasterProjectile(g_edict_t *ent, const vec3_t start, const vec3_t dir,
 		int32_t speed, int16_t damage, int16_t knockback);
 void G_GrenadeProjectile(g_edict_t *ent, const vec3_t start, const vec3_t dir,
